+++ mobile/components/profile/ProfilePage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+const ProfilePage = () => {
+  return (
+    <View>
+      <Text>Hello</Text>
+    </View>
+  );
+}
+
+export default ProfilePage
